Move packets::Unknown to sshwire::Unknown
This is emitted by sshwire-derive, so doesn't really belong in packets
module.
